+++ lib/src/upgrade_version.dart
@@ -1,9 +1,7 @@
 import 'dart:convert';
-import 'package:flutter/material.dart';
 import 'package:flutter_upgrade_version/models/package_info.dart';
 import 'package:flutter_upgrade_version/models/version_info.dart';
 import 'package:flutter_upgrade_version/src/util_service.dart';
-import 'package:html/parser.dart' show parse;
 import 'package:http/http.dart' as http;
 
 /// Method Type support
@@ -13,78 +11,15 @@ enum MethodType { postType, getType }
 ///
 abstract class UpgradeVersion {
   /// Get Upgrade version
-  /// Support Only Android
-  ///
-  @Deprecated(
-    'Use InAppUpdateManager for Android Device instead.',
-  )
-  static Future<VersionInfo?> getAndroidStoreVersion(
-      PackageInfo packageInfo) async {
-    final id = packageInfo.packageName;
-    final uri = Uri.https(
-        "play.google.com", "/store/apps/details", {"id": id, "hl": "en"});
-    try {
-      final response = await http.get(uri);
-      if (response.statusCode != 200) {
-        debugPrint('Can\'t find an app in the Play Store with the id: $id');
-        return VersionInfo(localVersion: packageInfo.version);
-      }
-      final document = parse(response.body);
-      String storeVersion = '0.0.0';
-      String? releaseNotes;
-
-      final additionalInfoElements = document.getElementsByClassName('hAyfc');
-      if (additionalInfoElements.isNotEmpty) {
-        final versionElement = additionalInfoElements.firstWhere(
-          (elm) => elm.querySelector('.BgcNfc')!.text == 'Current Version',
-        );
-        storeVersion = versionElement.querySelector('.htlgb')!.text;
-
-        final sectionElements = document.getElementsByClassName('W4P4ne');
-        final releaseNotesElement = sectionElements.firstWhere(
-          (elm) => elm.querySelector('.wSaTQd')!.text == 'What\'s New',
-        );
-        releaseNotes = releaseNotesElement
-            .querySelector('.PHBdkd')
-            ?.querySelector('.DWPxHb')
-            ?.text;
-      } else {
-        final scriptElements = document.getElementsByTagName('script');
-        final infoScriptElement = scriptElements.firstWhere(
-          (elm) => elm.text.contains('key: \'ds:5\''),
-        );
-
-        final param = infoScriptElement.text
-            .substring(20, infoScriptElement.text.length - 2)
-            .replaceAll('key:', '"key":')
-            .replaceAll('hash:', '"hash":')
-            .replaceAll('data:', '"data":')
-            .replaceAll('sideChannel:', '"sideChannel":')
-            .replaceAll('\'', '"');
-        final parsed = json.decode(param);
-        final data = parsed['data'];
-
-        storeVersion = data[1][2][140][0][0][0];
-        releaseNotes = data[1][2][144][1][1];
-      }
-
-      return VersionInfo(
-        localVersion: packageInfo.version,
-        storeVersion: storeVersion,
-        appStoreLink: uri.toString(),
-        releaseNotes: releaseNotes,
-      );
-    } catch (e) {
-      debugPrint('Exception: $e');
-      return VersionInfo(localVersion: packageInfo.version);
-    }
-  }
-
-  /// Get Upgrade version
   /// Support Only iOS
-  static Future<VersionInfo> getiOSStoreVersion(PackageInfo packageInfo) async {
+  //
+  // You need to pass country code as param if app is available on specific iTune store.
+  static Future<VersionInfo> getiOSStoreVersion({ required PackageInfo packageInfo, String? regionCode }) async {
     final id = packageInfo.packageName;
-    final parameters = {"bundleId": id};
+    final parameters = {
+      "bundleId": id,
+      "country" : regionCode,
+    };
     var uri = Uri.https("itunes.apple.com", "/lookup", parameters);
     Map<String, dynamic> _storeVersionGet =
         await getVersion(uri, MethodType.getType);